+++ src/it/unibs/fp/sistemaStellare/Satellite.java
@@ -4,7 +4,7 @@ public class Satellite extends CorpoCeleste{
     private int codicePianeta;
 
     //Costruttore luna
-    public Satellite (int codice, int massa, String nome, Posizione posizione, Pianeta pianeta) {
+    public Satellite (int codice, double massa, String nome, Posizione posizione, Pianeta pianeta) {
         //Parametri comuni al corpo celeste
         super(codice, massa, nome, posizione);
